+++ packages/liveblocks-core/src/room.ts
@@ -632,8 +632,6 @@ export type Room<
   getStorageStatus(): StorageStatus;
 
   /**
-   * @internal (for now)
-   *
    * Start an attempt to connect the room (aka "enter" it). Calling
    * `.connect()` only has an effect if the room is still in its idle initial
    * state, or the room was explicitly disconnected, or reconnection attempts
@@ -643,8 +641,6 @@ export type Room<
   connect(): void;
 
   /**
-   * @internal (for now)
-   *
    * Disconnect the room's connection to the Liveblocks server, if any. Puts
    * the room back into an idle state. It will not do anything until either
    * `.connect()` or `.reconnect()` is called.
@@ -822,11 +818,14 @@ export type RoomInitializers<
    */
   initialStorage?: TStorage | ((roomId: string) => TStorage);
   /**
-   * Whether or not the room connects to Liveblock servers. Default is true.
+   * Whether or not the room automatically connects to Liveblock servers.
+   * Default is true.
    *
    * Usually set to false when the client is used from the server to not call
    * the authentication endpoint or connect via WebSocket.
    */
+  autoConnect?: boolean;
+  /** @deprecated Renamed to `autoConnect` */
   shouldInitiallyConnect?: boolean;
 }>;
 
@@ -881,7 +880,7 @@ export function createRoom<
 >(
   options: Omit<
     RoomInitializers<TPresence, TStorage>,
-    "shouldInitiallyConnect"
+    "autoConnect" | "shouldInitiallyConnect"
   >,
   config: RoomConfig
 ): Room<TPresence, TStorage, TUserMeta, TRoomEvent> {